requirements.txt
Flask
pandas
nltk
openpyxl
python-dotenv
google-generativeai # Update this based on the actual version available

